move client to a splited package
