+++ expression/builtin_mvcc_test.go
@@ -1,13 +1,89 @@
 package expression_test
 
 import (
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 
-	"github.com/pingcap/check"
 	. "github.com/pingcap/check"
+	"github.com/pingcap/kvproto/pkg/kvrpcpb"
+	"github.com/pingcap/tidb/config"
+	"github.com/pingcap/tidb/server"
+	"github.com/pingcap/tidb/server/client"
+	"github.com/pingcap/tidb/util/mock"
 	"github.com/pingcap/tidb/util/testkit"
 )
+
+var _ = SerialSuites(&testServerSuite{})
+
+type testServerSuite struct {
+	testIntegrationSuiteBase
+
+	port       uint
+	statusPort uint
+
+	tidbdrv *server.TiDBDriver
+	tidbsrv *server.Server
+	tk      *testkit.TestKit
+	client  client.Client
+}
+
+// SetUpSuite create a mocked server mocked store
+func (s *testServerSuite) SetUpSuite(c *C) {
+	var err error
+	s.store, s.dom, err = newStoreWithBootstrap()
+	c.Assert(err, IsNil)
+	s.ctx = mock.NewContext()
+	s.tidbdrv = server.NewTiDBDriver(s.store)
+	globalConfig := config.GetGlobalConfig()
+	s.port = globalConfig.Port
+	s.statusPort = globalConfig.Status.StatusPort
+
+	cfg := config.NewConfig()
+	cfg.Port = s.port
+	cfg.Store = "tikv"
+	cfg.Status.StatusPort = s.statusPort
+	cfg.Status.ReportStatus = true
+
+	srv, err := server.NewServer(cfg, s.tidbdrv)
+	c.Assert(err, IsNil)
+	go srv.Run()
+
+	s.client = client.NewFromGlobalConfig()
+	err = s.client.PollServerOnline()
+	c.Assert(err, IsNil)
+
+	s.tk = testkit.NewTestKit(c, s.store)
+	s.tk.MustExec("create database tidb;")
+}
+
+// TearDownSuite will close mocked store and server
+func (s *testServerSuite) TearDownSuite(c *C) {
+	if s.dom != nil {
+		s.dom.Close()
+	}
+	if s.store != nil {
+		_ = s.store.Close()
+	}
+	if s.tidbsrv != nil {
+		s.tidbsrv.Close()
+	}
+}
+
+func (s *testServerSuite) prepare(c *C) {
+	s.tk.MustExec("use tidb;")
+}
+
+// TearDownTest will drop all table created by testing
+func (s *testServerSuite) clean(c *C) {
+	tk := s.tk
+	tk.MustExec("use tidb")
+	r := tk.MustQuery("show tables")
+	for _, tb := range r.Rows() {
+		tableName := tb[0]
+		tk.MustExec(fmt.Sprintf("drop table %v", tableName))
+	}
+}
 
 func (s *testServerSuite) TestMVCCGetInfo(c *C) {
 	s.prepare(c)
@@ -35,9 +111,9 @@ commit;
 
 	// load fixture to initialize test cases
 	fixture, err := s.getMVCCInfo("/mvcc/key/tidb/test/1")
-	c.Check(err, check.IsNil)
+	c.Check(err, IsNil)
 	var startTs uint64
-	if c.Check(len(fixture.Value.Info.Values), check.Greater, 0) {
+	if c.Check(len(fixture.Value.Info.Values), Greater, 0) {
 		startTs = fixture.Value.Info.Values[0].StartTs
 	}
 
@@ -62,10 +138,10 @@ commit;
 	}
 
 	for _, tt := range tests {
-		resp, err := s.fetchStatus(tt.Query)
-		c.Check(err, check.IsNil)
+		resp, err := s.client.Get(tt.Query)
+		c.Check(err, IsNil)
 		respBody, err := ioutil.ReadAll(resp.Body)
-		c.Check(err, check.IsNil)
+		c.Check(err, IsNil)
 		s.tk.MustQuery(fmt.Sprintf("select %s;", tt.Builtin)).Check(testkit.Rows(string(respBody)))
 	}
 }
@@ -95,9 +171,9 @@ commit;
 
 	// load fixture to initialize test cases
 	fixture, err := s.getMVCCInfo("/mvcc/key/tidb/pt(p2)/1")
-	c.Check(err, check.IsNil)
+	c.Check(err, IsNil)
 	var startTs uint64
-	if c.Check(len(fixture.Value.Info.Values), check.Greater, 0) {
+	if c.Check(len(fixture.Value.Info.Values), Greater, 0) {
 		startTs = fixture.Value.Info.Values[0].StartTs
 	}
 
@@ -116,10 +192,41 @@ commit;
 	}
 
 	for _, tt := range tests {
-		resp, err := s.fetchStatus(tt.Query)
-		c.Check(err, check.IsNil)
+		resp, err := s.client.Get(tt.Query)
+		c.Check(err, IsNil)
 		respBody, err := ioutil.ReadAll(resp.Body)
-		c.Check(err, check.IsNil)
+		c.Check(err, IsNil)
 		s.tk.MustQuery(fmt.Sprintf("select %s;", tt.Builtin)).Check(testkit.Rows(string(respBody)))
 	}
+}
+
+type mvccInfo struct {
+	Key      string                        `json:"key"`
+	RegionID uint64                        `json:"region_id"`
+	Value    *kvrpcpb.MvccGetByKeyResponse `json:"value"`
+}
+
+func (s *testServerSuite) getMVCCInfo(path string) (*mvccInfo, error) {
+	// retrieveMVCC information from http api
+	resp, err := s.client.Get(path)
+	if err != nil {
+		return nil, err
+	}
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	mvcc := &mvccInfo{}
+	if err := json.Unmarshal(body, &mvcc); err != nil {
+		return nil, err
+	}
+
+	// check rpc error
+	if mvcc.Value.Error != "" {
+		return nil, fmt.Errorf(mvcc.Value.Error)
+	}
+	if mvcc.Value.RegionError != nil {
+		return nil, fmt.Errorf(mvcc.Value.RegionError.Message)
+	}
+	return mvcc, nil
 }
